+++ src/mainboard/megaPi.js
@@ -1,12 +1,10 @@
 import Board from '../core/board';
 import electronics from '../electronic/index';
-import Mode from '../electronic/mode';
-import Control from '../core/control';
-import Version from '../electronic/version';
 
 const FIRMWAREMODE = Symbol('firmware');
 /**
  * MegaPi Class for 'MegaPi' mainboard.
+ * @private
  * @extends Board
  */
 class MegaPi extends Board{
@@ -19,8 +17,6 @@ class MegaPi extends Board{
     let this_ = this;
     //@member {String} {maiboard name}
     this.name = 'MegaPi';
-    //固件当前模式
-    this.currentMode = null;
     // 置空已连接块
     this.connecting = {};
     // 挂载电子模块
@@ -32,57 +28,6 @@ class MegaPi extends Board{
         };
       }
     }
-  }
-
-  /**
-   * set firmware mode
-   * @param {Number} mode 0、1、2、3、4
-   * @example
-   *   megapi.setFirmwareMode(1).run()
-   */
-  setFirmwareMode(mode){
-    this.currentMode = FIRMWAREMODE;
-    let subCmd = 0x12;
-    Mode.setMode(subCmd, mode);
-    return this;
-  }
-
-  /**
-   * get firmware mode
-   */
-  readFirmwareMode(){
-    this.currentMode = FIRMWAREMODE;
-    let subCmd = 0x72;
-    Mode.setMode(subCmd);
-    return this;
-  }
-
-  /**
-   * protocol that sent directly by mainboard firmware
-   * @return {Array}
-   */
-  get protocol (){
-    if(this.currentMode == FIRMWAREMODE) {
-      return Mode.protocol;
-    }else if( this.currentMode == 'version') {
-      return Version.protocol;
-    }
-  }
-
-  /**
-   * run
-   */
-  run() {
-    Control.write(Mode.protocol);
-    return this;
-  }
-
-  /**
-   * Get data of FirmwareMode
-   * @return {Promise}
-   */
-  async getData() {
-    return await Control.read(Mode.protocol);
   }
 }